+++ XuanYu.Editor.UI/Vm/Mode/UiVm.Mode.cs
@@ -36,6 +36,7 @@ public sealed partial class UiVm
         OnPropertyChanged(nameof(CurrentMode)); OnPropertyChanged(nameof(IsManageMode));
         OnPropertyChanged(nameof(IsEditMode)); OnPropertyChanged(nameof(IsMapEditMode));
         OnPropertyChanged(nameof(IsRegionEditMode)); OnPropertyChanged(nameof(CurrentEditorModeText));
+        OnPropertyChanged(nameof(CanStartRegionDrawing));
         OnPropertyChanged(nameof(IsMapEditorMode));
         RaiseLayerContextBindings();
     }